requirements.txt
discord.py>=2.3.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
